removing apikey from axios call, multiple room chat
